fix-bug="之前提交记录有敏感信息，移除敏感信息的几条提交记录后，提交记录变少了"

+++ OAuth-ResourceServer/src/main/java/com/wojucai/service/impl/ClientServiceImpl.java
@@ -1,13 +1,19 @@
 package com.wojucai.service.impl;
 
+import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.bean.copier.CopyOptions;
 import com.wojucai.dao.ClientRepository;
+import com.wojucai.dao.ConsentRepository;
 import com.wojucai.dao.ScopePropertyRepository;
 import com.wojucai.dao.ScopeRepository;
+import com.wojucai.entity.Bo.ConsentBo;
 import com.wojucai.entity.Bo.ScopeBo;
 import com.wojucai.entity.po.Client;
+import com.wojucai.entity.po.Consent;
 import com.wojucai.entity.po.Scope;
 import com.wojucai.entity.po.ScopeProperty;
 import com.wojucai.entity.reqParam.ClientQuery;
+import com.wojucai.entity.vo.ConsentVo;
 import com.wojucai.entity.vo.ScopeVo;
 import com.wojucai.service.ClientService;
 import com.wojucai.util.TextUtils;
@@ -20,9 +26,12 @@ import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import com.wojucai.entity.vo.ClientVo;
+import org.springframework.transaction.annotation.Transactional;
 
 import static com.wojucai.entity.codeEnum.ParamConstants.READ;
 import static com.wojucai.entity.codeEnum.ParamConstants.WRITE;
@@ -44,6 +53,8 @@ public class ClientServiceImpl extends AbstractImpl implements ClientService {
     private ScopePropertyRepository scopePropertyRepository;
     @Resource
     private ScopeRepository scopeRepository;
+    @Resource
+    private ConsentRepository consentRepository;
 
     @Override
     public Page<ClientVo> queryByClientName(ClientQuery clientQuery) {
@@ -102,13 +113,69 @@ public class ClientServiceImpl extends AbstractImpl implements ClientService {
             scopeVo.setScopeWrite(scopeWrite);
             res.add(scopeVo);
         }
-        System.out.println(res);
         return res;
     }
 
     @Override
+    public List<Integer> queryConsent(String clientId, Long userId) {
+        // 找到了对应的关联关系
+        Consent consent = consentRepository.findByClientIdAndAndUserId(clientId, userId);
+        if (consent == null) {
+            return new ArrayList<>();
+        }
+        return consent.getScope();
+    }
+
+    @Override
+    public List<ConsentVo> queryClientScope(String clientId) {
+        // 找到了对应的关联关系
+        Client client = clientRepository.findByClientId(clientId);
+        // 对关系进行分类
+        List<ScopeProperty> scopeProperties = scopePropertyRepository.findAllById(client.getScope());
+        // 获取对应的类和字段的对应关系
+        Map<Integer, List<ScopeProperty>> scopePropertyMap = scopeProperties.stream()
+                .collect(Collectors.groupingBy(ScopeProperty::getClassId));
+        List<ConsentVo> res = new ArrayList<>(scopePropertyMap.size());
+        scopePropertyMap.forEach((key,val)->{
+            System.out.println("类Id："+key+" ---属性集： "+val);
+            // 获取作用域
+            Optional<Scope> optional = scopeRepository.findById(key);
+            Scope scope = optional.get();
+            Map<String, List<ScopeProperty>> collect = val.stream().collect(Collectors.groupingBy(ScopeProperty::getDescription));
+            List<ConsentBo> consentBos = new ArrayList<>(2);
+            collect.forEach((description, propertyList) -> {
+                ConsentBo consentBo = new ConsentBo();
+                consentBo.setPropertyName(description);
+                propertyList.forEach(property -> {
+                    if (property.getBehavior() == 0) {
+                        consentBo.setReadId(property.getId());
+                    } else {
+                        consentBo.setWriteId(property.getId());
+                    }
+                });
+                consentBos.add(consentBo);
+            });
+            // 最后合为ConsentVo
+            ConsentVo consentVo = new ConsentVo(scope.getId(),scope.getScopeDescription(), consentBos);
+            res.add(consentVo);
+        });
+        return res;
+    }
+
+    @Override
+    public Client queryClientById(String id) {
+        return clientRepository.findByClientId(id);
+    }
+
+    @Transactional(rollbackFor = Exception.class)
+    @Override
     public void deleteById(Integer id) {
+        Client client = clientRepository.findById(id).orElse(null);
+        if (client == null) {
+            return;
+        }
         clientRepository.deleteById(id);
+        consentRepository.deleteByClientId(client.getClientId());
     }
 
     @Override
@@ -122,11 +189,20 @@ public class ClientServiceImpl extends AbstractImpl implements ClientService {
 
     @Override
     public Client updateClient(Client client) {
-        return clientRepository.save(client);
+        Client client1 = clientRepository.findById(client.getId()).get();
+        BeanUtil.copyProperties(client, client1, CopyOptions.create().ignoreNullValue());
+        return clientRepository.save(client1);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void batchDelete(List<Integer> ids) {
+        List<Client> allById = clientRepository.findAllById(ids);
         clientRepository.deleteAllByIdInBatch(ids);
+        allById.forEach(
+                client -> {
+                    consentRepository.deleteByClientId(client.getClientId());
+                }
+        );
     }
 }
